move test util into more appropriate place and get rid of circ imports
